Remove support for PHP < 7.2.24
We have previously used config/platform/php in composer.json to specify PHP versions.

We should still do this, but doing this alone would make PHP version bumps breaking changes (as the config option does not prevent you updating to a PHP version greater than that you're currently running via a `composer update` in a project that uses Iguana).

Therefore, I've also added the PHP version to `requires` - this means that people will only get this version of Iguana if they are running PHP 7.2.24 or greater, and therefore this is not a breaking change.