Add return type for PHP8 compatibility
Without these return types, we'll get deprecation notices when running
Iguana in PHP8, as they are required to comply with the PHP8 version of
the ArrayAccess interface.

The expected return type for `::offsetGet()` is `mixed`, but this does
not exist in PHP7. We can use the `#[\ReturnTypeWillChange]` attribute
to suppress the return type deprecation notice we'll otherwise get when
running PHP8 for now. It'll be treated as a comment in PHP7, see
https://www.php.net/manual/en/class.returntypewillchange.php and https://php.watch/versions/8.1/ReturnTypeWillChange